+++ app/scripts/statistics-grid.ts
@@ -9,7 +9,7 @@ import { Dataset, isTotalRow, Row, SingleRow } from "./statistics/statistics.typ
 import { StoreService } from "./services/store"
 import settings from "./settings"
 import { locObj } from "./i18n"
-import { formatFrequency } from "./util"
+import { escapeHtml, formatFrequency } from "./util"
 import { LangString } from "./i18n/types"
 import { zip } from "lodash"
 
@@ -109,8 +109,8 @@ function createColumns(store: StoreService, corpora: string[], attrs: string[]):
             field: "hit_value",
             sortable: true,
             formatter: (row, cell, value, columnDef, data: Row) => {
-                if (isTotalRow(data)) return "&Sigma;"
-                const output = data.formattedValue[reduceVal!] || `<span class="opacity-50">&empty;</span>`
+                if (isTotalRow(data)) return "Σ"
+                const output = escapeHtml(data.formattedValue[reduceVal!]) || `<span class="opacity-50">&empty;</span>`
                 return `<div class="link" data-row="${data.rowId}" ${dir}>${output}</div>`
             },
             minWidth,